process/pyproject.toml
[project]
name = "process"
version = "0.1.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "netcdf4>=1.7.4",
    "numpy>=2.4.0",
    "pillow>=12.1.0",
    "psycopg2-binary>=2.9.7",
    "pyproj>=3.7.2",
    "requests>=2.32.5",
    "scipy>=1.16.3",
    "xarray>=2025.12.0",
]
